requirements.txt
Django==3.2.3
djangorestframework==3.12.4

# tests and debug
factory-boy==3.2.0
ipdb==0.13.7
pytest==6.2.4
pytest-cov==2.12.0
pytest-django==4.3.0
pytest-sugar==0.9.4
django-coverage-plugin==1.8.0
django-extensions==3.1.3

# style
black==21.5b1
flake8==3.9.2
pylint==2.8.2
pylint-common==0.2.5
pylint-django==2.4.4
pylint-plugin-utils==0.6
